refactor: changes repositories of db queries names
