+++ src/routes/user-routes.ts
@@ -1,5 +1,5 @@
 import { UsersController } from "@/adapters/controllers/users-controller";
-import { MethodsUsersRepository } from "@/adapters/repositories/users-repository";
+import { MethodsUsersRepository } from "@/adapters/repositories/db-users-repository";
 import { CreateUserUseCase } from "@/core/use-cases/users/create-user";
 import type { FastifyInstance } from "fastify";
